pyproject.toml
[tool.poetry]
name = "grizzly"
version = "0.1.2"
description = "MAD Consulting - Grizzly"
authors = ["victor-vila <victor.vila.navazo@gmail.com>"]
packages = [{include = "grizzly_main"}]

# TODO - In the future we can use multiple dependency groups / extras, if we start adding lots of dependencies
# which might not be required for all the subprojects within the grizzly_main repo
[tool.poetry.dependencies]
python = ">=3.9, <=3.11"
pytest = "7.4.0"
# Spark local dependencies
pyspark = "3.3.0"
numpy = "1.23.4"
docker-compose = "1.29.2"
jupyterlab = "3.4.8"
# Pulumi dependencies
pulumi = ">=3.0.0,<4.0.0"
pulumi-aws = ">=5.0.0,<6.0.0"
# Spark EMR Serverless dependencies
botocore = "1.29.13"
boto3 = "1.26.12"
tomlkit = "0.11.6"
typer = "0.7.0"
rich = "13.3.1"
"ruamel.yaml" = "0.17.21"
seedir =  "0.4.2"
flask = "2.3.3"
fastapi = "0.100.1"

# Optional dependency group for the Prefect POC (in grizzly_main/orchestrator/prefect_poc)
[tool.poetry.group.prefect_poc]
optional = true

[tool.poetry.group.prefect_poc.dependencies]
prefect = "2.13.5"
pysqlite3 = "0.5.1"
mysql-connector-python = "8.1.0"
pandas = "2.1.0"
featuretools = "1.27.0"
matplotlib = "3.5.2"
lightgbm = "4.1.0"
optuna = "3.3.0"
mysqlclient = "2.2.0"
pyarrow = "13.0.0"
mlflow = "2.7.1"
plotly = "5.17.0"
kaleido = "0.2.1"
shap = "0.42.1"
ipython = "8.16.1"
fasttreeshap = "0.1.6"


[tool.poetry.group.dev.dependencies]
pre-commit = "3.3.3"
# Adding pre-commit hooks libraries to be available to be used independently via the Poetry environment
black = "23.7.0"
blacken-docs = "1.16.0"
isort = "5.11.2"
bandit = "1.7.5"
mypy = "1.5.1"
flake8 = "6.1.0"
pylint = "2.17.5"
interrogate = "1.5.0"
pdoc3 = "0.10.0"
pytest-cov = "4.1.0"
codespell = "2.2.5"

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120

[tool.isort]
line_length = 120

[tool.pylint.format]
max-line-length = 120

[tool.codespell]
skip = '*.lock'

[build-system]
requires = ["poetry-core>=1.2.0b3"]
build-backend = "poetry.core.masonry.api"
